Build and test cosmopolitan and uurl
The HTTP message parser from cosmopolitan libc has been copied here
with the least amount of modifications as possible. A static library
is being built and ran on unit tests that match those of the original
repo plus some others.

A second static library, uurl, is also being built and ran on the same
tests.

The goal is to adopt the parser from cosmopolitan for uurl's purposes
but to refactor it to my own personal standards. It's important that any
changes to uurl pass the same tests as cosmopolitan's parser to keep them
in sync because parsing strings in C is hard and I'm not trying to pretend
that I'm good at it.